restructure the vagrantfile to use the vmware instead

+++ p1/scripts/kubectl.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
+set -e
+
 sudo apt-get update && sudo apt-get upgrade -y
-curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
-chmod +x kubectl
-sudo mv kubectl /usr/local/bin/
+
+if [ ! -f /usr/local/bin/kubectl ]; then
+  echo "[INFO] Downloading kubectl..."
+  KUBECTL_VERSION=$(curl -s https://dl.k8s.io/release/stable.txt)
+  curl -L --retry 5 --retry-delay 10 --output kubectl "https://dl.k8s.io/release/${KUBECTL_VERSION}/bin/linux/arm64/kubectl"
+  chmod +x kubectl
+  sudo mv kubectl /usr/local/bin/
+else
+  echo "[INFO] kubectl already exists."
+fi
